+++ examples/components/welcome.js
@@ -2,15 +2,18 @@ import {Layout} from './layout';
 import {Link} from 'react-router';
 import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
+import {prefixAll} from 'inline-style-prefixer';
+
+const style = prefixAll({
+  display: '-webkit-flex;flex',
+  justifyContent: 'space-around',
+  width: '100vmin'
+});
 
 function WelcomeComponent (props) {
   return (
     <Layout>
-      <div style={{
-        display: 'flex',
-        justifyContent: 'space-around',
-        width: '100vmin'
-      }}>
+      <div style={style}>
       {props.components.map(({key, name}) => {
         return <Link key={key} to={`/${key}`}>{name}</Link>;
       })}